requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
yfinance>=0.2.18
tiktoken>=0.4.0
tenacity>=8.2.0
